Add wheel building workflow and packaging improvements
A docker image is made in necessary129/manylinux2014-LLVM with
prebuilt LLVM + cmake + ninja

cibuildwheel is being used to generate the wheels. Which
uses `auditwheel` to bundle linked libraries. So the ELF
patching in `preparation.py` is not needed.

The includes can be be found by Python by itself. Only `utils.h`
needs to be specifically included.